+++ squash_wip.go
@@ -13,7 +13,7 @@ type Replacer func(string) string
 func squashWip(configuration Configuration) {
 	if !isMobProgramming(configuration) {
 		sayError("you aren't mob programming")
-		sayTodo("to start mob programming, use", "mob start")
+		sayTodo("to start, use", "mob start")
 		return
 	}
 
@@ -26,7 +26,7 @@ last commit must be a manual commit`)
 	}
 
 	currentBaseBranch, currentWipBranch := determineBranches(gitCurrentBranch(), gitBranches(), configuration)
-	mergeBase := silentgit("merge-base", currentWipBranch, currentBaseBranch)
+	mergeBase := silentgit("merge-base", currentWipBranch.String(), currentBaseBranch.String())
 
 	originalGitEditor, originalGitSequenceEditor := getEnvGitEditor()
 	setEnvGitEditor(
@@ -35,9 +35,9 @@ last commit must be a manual commit`)
 	)
 	silentgit("rebase", "-i", "--keep-empty", mergeBase)
 	setEnvGitEditor(originalGitEditor, originalGitSequenceEditor)
-	sayInfo("the history of your " + currentWipBranch + " branch has been rewritten to combine all wip commits with their following manual commits:")
+	sayInfo("the history of your '" + currentWipBranch.String() + "' branch has been rewritten to combine all wip commits with their following manual commits:")
 	sayEmptyLine()
-	sayLastCommitsWithMessage(currentBaseBranch, currentWipBranch)
+	sayLastCommitsWithMessage(currentBaseBranch.String(), currentWipBranch.String())
 	sayEmptyLine()
 	sayTodo("to finally put the changes into the base branch preserving the resulting commits, call:", "mob done --no-squash")
 }
@@ -47,7 +47,7 @@ func sayLastCommitsWithMessage(currentBaseBranch string, currentWipBranch string
 	log := silentgit("--no-pager", "log", commitsBaseWipBranch, "--pretty=oneline", "--abbrev-commit")
 	lines := strings.Split(log, "\n")
 	if len(lines) > 10 {
-		sayInfo("This mob branch contains " + strconv.Itoa(len(lines)) + " commits. The last 10 were:")
+		sayInfo("wip branch '" + currentWipBranch + "' contains " + strconv.Itoa(len(lines)) + " commits. The last 10 were:")
 		lines = lines[:10]
 	}
 	output := strings.Join(lines, "\n")
@@ -95,7 +95,7 @@ func squashWipGitSequenceEditor(fileName string, configuration Configuration) {
 }
 
 func replaceFileContents(fileName string, replacer Replacer) {
-	file, err := os.OpenFile(fileName, os.O_RDWR, 0666)
+	file, _ := os.OpenFile(fileName, os.O_RDWR, 0666)
 	input, err := ioutil.ReadAll(file)
 	if err != nil {
 		panic(err)
@@ -127,7 +127,7 @@ func endsWithWipCommit(configuration Configuration) bool {
 
 func commitsOnCurrentBranch(configuration Configuration) []string {
 	currentBaseBranch, currentWipBranch := determineBranches(gitCurrentBranch(), gitBranches(), configuration)
-	commitsBaseWipBranch := currentBaseBranch + ".." + currentWipBranch
+	commitsBaseWipBranch := currentBaseBranch.String() + ".." + currentWipBranch.String()
 	log := silentgit("--no-pager", "log", commitsBaseWipBranch, "--pretty=format:%s")
 	lines := strings.Split(log, "\n")
 	return lines